split library_test_utils.go to avoid linter warnings
